Moved render and universe folders out of engine folder into src folders and updated CMakeLists.txt files. Found a prespective problem by moving object within 1.0 because it still uses frustum defaults. Working in progress. Replaced sphere generator and now displays sphere.

+++ src/engine/player.h
@@ -16,17 +16,17 @@ public:
 
 	inline vec3d_t position() { return camPosition; }
 	inline quatd_t rotation() { return camRotation; }
-	inline float getFOV() { return fov; };
+	inline double getFOV() { return fov; };
 
-	void setFOVrad(float fov);
-	void setFOVdeg(float fov);
+	void setFOVrad(double fov);
+	void setFOVdeg(double fov);
 
 	void focus(Object *obj);
 
 private:
 	vec3d_t camPosition;
 	quatd_t camRotation;
-	float   fov;
+	double  fov;
 };
 
 class Player {
